pyproject.toml
[project]
name = "ocr-documents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "elasticsearch>=8.17.2",
    "fastapi>=0.115.12",
    "pillow>=11.1.0",
    "pymongo>=4.12.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.0",
]
